Allow cancelling broken citation edit without removing it from the doc
Before we had the Please Wait... screen on google docs we used to
have to immediatelly insert the placeholder for new citations
so that if the user moves the cursor while citation dialog is open
we still know where the insert should be made.

The side effect of this is that broken citations are treated as
not citations at all and a new placeholder needs to be reinserted.

We now delay insertion of the placeholder until we start pushing
doc changes via gdocs AP